fix: harden internal functions against out-of-contract inputs
- Makefile: add -DFUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION to FUZZ_CFLAGS
  so standalone main() is excluded when building with clang -fsanitize=fuzzer.
  Older clang versions (<=10) don't define this macro automatically.
- fuzz_stats.c: bound n against available data in ow_stats_capture_pulse;
  defensive hardening against out-of-range caller (flagged by ASan under
  fuzzing)
- ds18b20.c: cast res_config_byte left-shift to uint8_t to avoid UBSan
  flagging left-shift-of-negative for res < DS18B20_RES_MIN

+++ src/ds18b20.c
@@ -491,7 +491,7 @@ uint8_t ds18b20_alarm_search_count(void) { return onewire_search_count(); }
  * @note 9 bit -> 0x1F, 10 bit -> 0x3F, 11 bit -> 0x5F, 12 bit -> 0x7F.
  */
 __STATIC_FORCEINLINE uint8_t res_config_byte(uint8_t res) {
-    return (uint8_t)(0x1Fu | ((res - DS18B20_RES_MIN) << 5));
+    return (uint8_t)(0x1Fu | ((uint8_t)(res - DS18B20_RES_MIN) << 5));
 }
 
 /**
